+++ apps/landing/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @fluid-ds/landing
 
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @fluid-ds/components@0.1.5
+  - @fluid-ds/parser@0.1.5
+
 ## 0.0.5
 
 ### Patch Changes